Move Kubelet pod-management code into kubelet_pods.go
